Potential core plugin approach
During the build process, the lines in between "BEGIN EXCISION" and "END
EXCISION" comments would be split out to the file indicated in the
comments. In the release builds, that section of code would be replaced
with a `require` statement that only takes effect when the given "core
plugin" is active.